requirements.txt
PyYAML
aiohttp
aiofiles
pandas
cerberus